I have created an alert(popup) when you delete a selected appointment

+++ app/src/containers/Appointment/Appointment.jsx
@@ -47,12 +47,17 @@ const Appointment = () => {
                 'Authorization': 'Bearer ' + token
             }
         }
+        const doAppointment = window.confirm('You are about to delete your appointment, are you sure?')
+        if(doAppointment == true){
+            console.log('Hemos entrado')
+            const appointmentData = await axios.delete(URL + `${argument.id}`, config)
+            console.log(appointmentData);
+
+            return setTimeout(() => {
+                history.push('/schedule')
+            }, 1000);
+        }
         /*`http://localhost:3000/appointment/${argument.id}`*/
-        const appointmentData = await axios.delete(URL + `${argument.id}`, config)
-        console.log(appointmentData);
-        return setTimeout(() => {
-            history.push('/schedule')
-        }, 1000);
 
     }
